code cleanup and refactoring for the stretch codebase
