+++ src/home_robot/home_robot/agent/ovmm_agent/__init__.py
@@ -8,6 +8,7 @@ from .ovmm_agent import OpenVocabManipAgent, Skill, get_skill_as_one_hot_dict
 from .ovmm_perception import (
     OvmmPerception,
     build_vocab_from_category_map,
+    create_semantic_sensor,
     read_category_map_file,
 )
 from .pick_and_place_agent import PickAndPlaceAgent
