+++ src/home_robot/home_robot/utils/config.py
@@ -68,3 +68,38 @@ def get_control_config(cfg_name):
         cfg = hydra.compose(config_name=cfg_name)
 
     return cfg
+
+
+def load_config(
+    visualize: bool = False, print_images: bool = True, config_path=None, **kwargs
+):
+    """Load config path for real world experiments and use proper presets."""
+    if config_path is None:
+        # TODO: make sure this is the right default
+        config_path = "projects/real_world_ovmm/configs/agent/eval.yaml"
+    config, config_str = get_config(config_path)
+    config.defrost()
+    config.NUM_ENVIRONMENTS = 1
+    config.VISUALIZE = int(visualize)
+    config.PRINT_IMAGES = int(print_images)
+    config.EXP_NAME = "debug"
+    if config.GROUND_TRUTH_SEMANTICS != 0:
+        raise RuntimeError("No ground truth semantics in the real world!")
+    config.freeze()
+    return config
+
+
+def load_slap_config(
+    visualize: bool = False, print_images: bool = True, config_path=None, **kwargs
+):
+    """Load config path for real world experiments with SLAP manipualtion policies; configure to use proper presets."""
+    config, config_str = get_config(config_path)
+    config.defrost()
+    config.NUM_ENVIRONMENTS = 1
+    config.VISUALIZE = int(visualize)
+    config.PRINT_IMAGES = int(print_images)
+    config.EXP_NAME = "debug"
+    if config.GROUND_TRUTH_SEMANTICS != 0:
+        raise RuntimeError("No ground truth semantics in the real world!")
+    config.freeze()
+    return config
